initial clean up
