+++ App.tsx
@@ -1,8 +1,8 @@
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { SafeAreaView, ScrollView, StyleSheet, Text, View,Button } from 'react-native';
-import Home from './screens/home';
-import deviceDetails from './screens/deviceDetails';
+import { StyleSheet } from 'react-native';
+import Home from './screens/Home';
+import DeviceDetails from './screens/DeviceDetails';
 
 const Stack = createNativeStackNavigator()
 
@@ -16,61 +16,9 @@ export default function App() {
         />
         <Stack.Screen
           name="Details"
-          component={deviceDetails}
+          component={DeviceDetails}
         />
       </Stack.Navigator>
-
     </NavigationContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  button: { 
-    justifyContent: 'flex-end',
-  },
-  devicesView:{
-    flex:1,
-    marginVertical:30,
-    marginHorizontal:15,
-  },
-  scrollView:{
-    marginVertical:15,
-  },
-  device:{
-    borderColor:'green',
-    borderRadius: 5,
-    borderWidth: 0.4,
-    padding:7,
-    marginVertical:5,
-    flexDirection: "row"
-  },
-  titles: {
-    marginTop: 30,
-    alignItems: 'center',
-    color:'darkgreen',
-  },
- 
-  title: {
-    color:'darkgreen',
-    fontSize: 60,
-    fontWeight: '500',
-  },
-  subtitle: {
-    color:'darkgreen',
-  },
-  
-  mainLayout: {
-    height:"100%",
-    // backgroundColor: 'lightgreen',
-  },
-  statusbar: {
-    height:50,
-    backgroundColor: 'darkgreen',
-  },
-});
