+++ screens/DeviceDetails/index.tsx
@@ -1,18 +1,16 @@
 import { StatusBar } from 'expo-status-bar';
 import { SafeAreaView, ScrollView, StyleSheet, Text, View,Button,Image } from 'react-native';
-// import deviceIcon from '../assets/deviceIcon.png'
+import deviceIcon from '../../assets/favicon.png'
 
-export default function deviceDetails() {
+export default function DeviceDetails() {
   return (
     <View style={styles.mainLayout}>
-        <StatusBar backgroundColor='darkgreen'/>
-      
-      
+      <StatusBar backgroundColor='darkgreen'/>
       <SafeAreaView style={styles.titles}>
-        {/* <Image
+        <Image
             style={{height:100,width:100}}
             source={deviceIcon}
-        /> */}
+        />
         <Text style={{fontWeight:'700',color:'darkgreen'}}>Bedroom device</Text>
       </SafeAreaView>
       <View style={styles.devicesView}>
@@ -54,12 +52,6 @@ export default function deviceDetails() {
 }
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
   button: { 
     justifyContent: 'flex-end',
   },
@@ -84,22 +76,8 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     color:'darkblue',
   },
- 
-  title: {
-    color:'darkblue',
-    fontSize: 60,
-    fontWeight: '500',
-  },
-  subtitle: {
-    color:'darkblue',
-  },
   
   mainLayout: {
     height:"100%",
-    // backgroundColor: 'lightblue',
-  },
-  statusbar: {
-    height:50,
-    backgroundColor: 'darkblue',
   },
 });
