+++ screens/Home/index.tsx
@@ -1,24 +1,12 @@
 import { StatusBar } from 'expo-status-bar';
-import { SafeAreaView, ScrollView,Image, StyleSheet, Text, View,Button,TouchableOpacity } from 'react-native';
-import deviceIcon from '../assets/favicon.png'
+import { SafeAreaView, ScrollView, StyleSheet, Text, View, Button } from 'react-native';
+import Device from './components/Device'
 
-const Device = ({name,connected,navigation}:{name:string,connected:boolean,navigation:any}) =>{
-  return(
-    <TouchableOpacity style={styles.device} onPress={()=>navigation.navigate('Details')} >
-        <Image
-            style={{height:40,width:40}}
-            source={deviceIcon}
-        />
-      <View>
-        <Text style={{fontWeight:'700'}}>{name}</Text>
-        <Text>{connected? 'Connected' : 'Not connected'}</Text>
-      </View>
-      
-    </TouchableOpacity>
-  )
+type HomeProps = {
+  navigation:any
 }
 
-export default function Home({ navigation }) {
+export default function Home({ navigation }:HomeProps) {
   return (
     <View style={styles.mainLayout}>
       <StatusBar backgroundColor='darkgreen'/>
@@ -27,6 +15,7 @@ export default function Home({ navigation }) {
         <Text style={styles.title}>WHOSS</Text>
         <Text style={styles.subtitle}>Wireless home occupancy and security system</Text>
       </SafeAreaView>
+
       <View style={styles.devicesView}>
         <Text style={{fontWeight:'600',color:'darkgreen'}}>Devices</Text>
         <ScrollView style={styles.scrollView}>
@@ -34,6 +23,7 @@ export default function Home({ navigation }) {
           <Device name="Living room device" connected={false} navigation={navigation}/>
           <Device name="Hallway device" connected={false} navigation={navigation}/>
         </ScrollView>
+
         <View style={styles.button}>
           <Button
           onPress={()=>navigation.navigate('Details')} 
@@ -49,12 +39,6 @@ export default function Home({ navigation }) {
 }
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
   button: { 
     justifyContent: 'flex-end',
   },
@@ -65,14 +49,6 @@ const styles = StyleSheet.create({
   },
   scrollView:{
     marginVertical:5,
-  },
-  device:{
-    borderColor:'green',
-    borderRadius: 5,
-    borderWidth: 0.4,
-    padding:7,
-    marginVertical:5,
-    flexDirection: "row"
   },
   titles: {
     marginTop: 30,
@@ -88,13 +64,7 @@ const styles = StyleSheet.create({
   subtitle: {
     color:'darkgreen',
   },
-  
   mainLayout: {
     height:"100%",
-    // backgroundColor: 'lightgreen',
-  },
-  statusbar: {
-    height:50,
-    backgroundColor: 'darkgreen',
   },
 });
